Updating to make rfitoolbox more current with the SAM implementation
