+++ rfi_toolbox/datasets/__init__.py
@@ -5,5 +5,6 @@ Provides batched streaming datasets for memory-efficient training.
 """
 
 from .batched_dataset import TorchDataset, BatchWriter
+from .rfi_mask_dataset import RFIMaskDataset
 
-__all__ = ["TorchDataset", "BatchWriter"]
+__all__ = ["TorchDataset", "BatchWriter", "RFIMaskDataset"]
