+++ rfi_toolbox/visualization/__init__.py
@@ -1,2 +1,11 @@
-# rfitool_box/visualization/__init__.py
-# This module provides visualization utilities for the RFI Toolbox.
+"""
+Interactive visualization utilities for RFI data.
+
+Provides Bokeh-based interactive viewers for exploring RFI datasets
+and model predictions.
+
+Example:
+    $ visualize_rfi_data --dataset_dir ./data/val --model_path model.pt
+"""
+
+__all__ = []
